dracut: enable swap as early as possible
Enable swap just after creating the partition. This ensures early
startup services (including fsck which in recent Fedora versions is
called inside initramfs already) will not run out of memory.

Do this only for the in-VM kernel case (initramfs built by the VM
itself). Avoid this change in "simple" dracut module (for dom0-provided
kernel), as it may result in surprising effects in the VM that didn't
expected it at this stage (for example when it calls mkswap itself).

Fixes QubesOS/qubes-issues#6174