Use "opened" Xenstore key instead of magic file in /etc
Changing behavior based on whether or not a magic file exists in /etc is
a bad idea.  Instead, have the kernel create an "opened" key (with the
value 0) to indicate that it supports autocleared devices and diskseqs
in Xenstore, and use the existence of that key to determine behavior.

NOT TESTED.