Generate allowlist at compile-time
This has four significant advantages:

- It avoids a runtime dependency on the ICU-C library and its associated
  data files.

- It ensures that the allowlist is independent of the version of ICU-C
  provided by the distribution, without having to bundle ICU-C with
  libqubes-pure.

- Using ICU-C is expected to be much slower than a raw C switch()
  statement optimized by GCC or Clang.

- It is possible for the generated allowlist to be inspected and
  audited.

The generated allowlist is checked into git to avoid build-time
dependencies on the correct version of ICU-C.  To prevent the generated
allowlist from becoming out-of-date, building for Fedora 36 checks that
the allowlist produced by the generator is the same as the allowlist in
the git repository.  If the allowlists differ, the build fails in
%check.  Fedora 36 was chosen for the simple reason that it is the
distribution I (Demi Marie Obenour) was using when making this commit,
and therefore the distribution used to generate
unicode-allowlist-table.c.