initramfs: Handle the case where device-mapper is a module
Qubes OS-provided kernels have device-mapper builtin, but it can also be
built as a module, in which case /dev/mapper will not exist until the
module is loaded or userspace creates the directory.  If the root volume
is writable, this causes mounting /sysroot to fail, which in turn causes
switch_root to fail with a confusing error.  This causes Linux to panic
with "Attempted to kill init!".  Avoid this problems by explicitly
creating /dev/mapper if it does not already exist.

Fixes: QubesOS/qubes-issues#8220