Stricter symbolic link checks
A copied symbolic link should not be able to escape the directory tree
it is in.  This means that top-level symbolic links must be rejected,
and that a link named a/b/c must be allowed to point to ../b (which
would resolve to a/b) but not ../../b (which would resole to b).  This
ensures that users can write to directory trees copied by qvm-copy
without worrying about a symlink attack causing them to write to a
different location in their filesystem, even if the directory tree has
been moved out of ~/QubesIncoming/VMNAME.

It turns out that a/b pointing to .. or ../a was rejected, but top-level
symbolic links were not.  Fix that, add unit tests, and make the code
far, *far* easier to understand.