Avoid writing junk into Xenstore
The path buffer pointer would be overwritten, but the length would not
be.  xs_write() would therefore read out of bounds.

This is not a security problem as the not-script process has no secrets
in its address space, and besides it is not exposed to untrusted input.

Not tested beyond "it builds", but should be quite obvious.

Fixes: QubesOS/qubes-issues#8708