Do not use fchmodat() because it follows symlinks
This is not a security vulnerability, since an attacker cannot cause a
symbolic link to replace a file and the symlink restrictions mean that
the attacker could only change permissions of paths inside
~/QubesIncoming/VMNAME/TOPLEVEL_DIR anyway.  Still, not following
symbolic links is the right thing to do.  With pre-6.6 Linux kernels,
fchmodat(fd, path, AT_SYMLINK_NOFOLLOW) is broken in a chroot without
/proc mounted, so it cannot be used.  Instead, open the path and call
fchmod() on the file descriptor.

Fixes: ed68c01716b4 ("Use FD-based versions chmod and utime")